serialization: move serialization into separate module
